refactor: move panelview to workspace/panels
