+++ packages/base/src/workspace/panels/components/legendItem.tsx
