+++ packages/base/src/workspace/panels/index.ts
